requirements.txt
asgiref==3.8.1
Django==4.2
django-admin-interface==0.28.6
django-colorfield==0.11.0
djangorestframework==3.15.1
gunicorn==22.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
whitenoise==6.6.0
